vtable: allow resetting column alignment when table data changes
* lisp/emacs-lisp/vtable.el (vtable--compute-columns): If a column
was not created with an explicit 'align' property, allow changing
this property when the column data changes from numeric to
non-numeric (or vice versa).  This makes it possible to add data to
an empty table, because in a table without data all columns are
assumed to be numeric and right-aligned.  (Bug#73775)