Remove support for the `editorconfig` executable
Some of the code lacks copyright paperwork, so let's remove it for now,
so we don't have to adapt it to the new hooks.
If/when it proves useful/necessary it'll be easy enough to re-add.

* editorconfig.el (editorconfig-exec-path, edconf-get-properties-function)
(edconf-exec-path, editorconfig-get-properties-function): Delete variables.
(editorconfig--execute-editorconfig-exec)
(editorconfig--parse-properties, editorconfig-get-properties)
(editorconfig-get-properties-from-exec): Delete functions.
(editorconfig-call-get-properties-function):
Call `editorconfig-core-get-properties-hash` directly.