Prevent which-key faces from over-extending during formatting
* lisp/which-key.el (which-key--pad-column): Instead of
injecting the description using a left-aligned format-string (%-[n]s),
we manually add the necessary padding so that the text
properties don't extend beyond the description.  This allows us
to use a face with a background without it extending until the
end of a column.

See also 63a6fb2a7a02ca88835c3fd473894d3b7d39ff15 for a similar
issue with Quick Help.